feat: - refactor squidexCollections function for improved error handling and schema loading
- refactor Squidex client and schema handling for improved organization and clarity
