+++ packages/starsquid/src/data/core/api.ts
@@ -2,7 +2,6 @@
 import { SquidexClient } from "@squidex/squidex";
 import { InMemoryTokenStore } from "@squidex/squidex/dist/wrapper/SquidexClient.js";
 
-
 export function SquidexClientFactory(
   squidexAppName: string,
   squidexClientId: string,
